# internal/api/openapi.json
{
  "openapi": "3.2.1",
  "info": {
    "title": "WG-Guard API",
...
    "version": "1.0.0"
  },
  "servers": [{ "url": "/" }],
  "security": [{ "BearerAuth": [] }],
...
  "paths": {
...
    "/api/v1/interfaces": {
      "get": {
        "tags": ["interfaces"], "summary": "List tunnel interfaces/profiles", "security": [{ "BearerAuth": ["interfaces.read"] }],
        "responses": { "200": { "description": "Interfaces", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Interface" } } } } } } } }
      },
      "post": {
        "tags": ["interfaces"], "summary": "Create a profile (server keypair generated)", "security": [{ "BearerAuth": ["interfaces.write"] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string", "pattern": "^awg[0-9]+$", "example": "awg0" }, "listen_port": { "type": "integer", "description": "UDP port number, not a byte or time quantity." }, "ipv4_subnet": { "type": "string", "example": "10.8.0.0/24", "description": "Primary canonical RFC1918 IPv4 CIDR (/29 or larger). /23, /22, /21 and /20 are supported. Mutually exclusive with ipv4_pools. Omit both for an available default pool." }, "ipv4_pools": { "$ref": "#/components/schemas/IPv4Pools" }, "mtu": { "type": "integer", "description": "Tunnel IP-packet MTU in bytes (B); not a traffic allowance." }, "preset": { "type": "string", "enum": ["plain", "recommended", "performance", "balanced", "resilient", "suggested", "randomized"], "description": "Server-owned profile policy. Omit this field when sending explicit obfuscation; preset and obfuscation are mutually exclusive." }, "backend_mode": { "type": "string", "enum": ["kernel", "userspace"] }, "endpoint_override": { "type": "string", "description": "Canonical hostname or IP address, optionally with port; URLs, paths, whitespace, and control characters are rejected.", "example": "vpn.example.com:51820" }, "obfuscation": { "$ref": "#/components/schemas/Obfuscation" } } } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Interface" } } } }, "409": { "$ref": "#/components/responses/Conflict" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer", "description": "API token: `Authorization: Bearer wg_…` with scopes; separate from admin sessions." }
    },
...
    "responses": {
...
      "Conflict": { "description": "Conflict (exists, limit reached, pool exhausted, in use)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "request_id"],
            "properties": {
              "code": { "type": "string", "description": "Stable machine code: USER_NOT_FOUND, USERNAME_EXISTS, DEVICE_LIMIT_REACHED, TRAFFIC_EXCEEDED, INVALID_REQUEST, UNAUTHORIZED, FORBIDDEN, RATE_LIMITED, NODE_UNAVAILABLE, INTERNAL_ERROR, …" },
              "message": { "type": "string" },
              "request_id": { "type": "string" }
            }
          }
        }
      },
...
      "Obfuscation": {
        "type": "object",
        "additionalProperties": false,
        "description": "Complete supported AmneziaWG profile at the pinned revisions. Scalar-or-range values are inclusive: a scalar is a JSON integer and a true range is canonical string N-M. Enabled H1-H4 intervals must be non-zero and pairwise non-overlapping. AdvancedSecurity is intentionally unsupported.",
        "properties": {
          "enabled": { "type": "boolean" },
          "jc": { "type": "integer", "minimum": 0, "maximum": 128, "description": "Junk packet count; enabled profiles require 1-128.", "example": 5 },
          "jmin": { "type": "integer", "minimum": 0, "maximum": 1279, "description": "Minimum junk size; must be less than jmax.", "example": 40 },
          "jmax": { "type": "integer", "minimum": 0, "maximum": 1280, "description": "Maximum junk size; must exceed jmin.", "example": 70 },
          "s1": { "type": "integer", "minimum": 0, "maximum": 1132, "description": "Handshake-init padding size.", "example": 86 },
          "s2": { "type": "integer", "minimum": 0, "maximum": 1188, "description": "Handshake-response padding size; s1 + 56 must not equal s2.", "example": 61 },
          "s3": { "type": "integer", "minimum": 0, "maximum": 65535, "description": "Cookie-message padding size; explicit advanced option.", "example": 40 },
          "s4": { "type": "integer", "minimum": 0, "maximum": 65535, "description": "Transport-message padding size; explicit advanced option.", "example": 48 },
          "h1": { "description": "Inclusive u32 magic-header scalar or range.", "example": "100-110", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 4294967295 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "h2": { "description": "Inclusive u32 magic-header scalar or range.", "example": 200, "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 4294967295 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "h3": { "description": "Inclusive u32 magic-header scalar or range.", "example": "300-310", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 4294967295 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "h4": { "description": "Inclusive u32 magic-header scalar or range.", "example": 400, "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 4294967295 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "i1": { "type": "string", "description": "Canonical single-line UTF-8 signature value; leading/trailing whitespace and control characters are rejected." },
          "i2": { "type": "string", "description": "Canonical single-line UTF-8 signature value; leading/trailing whitespace and control characters are rejected." },
          "i3": { "type": "string", "description": "Canonical single-line UTF-8 signature value; leading/trailing whitespace and control characters are rejected." },
          "i4": { "type": "string", "description": "Canonical single-line UTF-8 signature value; leading/trailing whitespace and control characters are rejected." },
          "i5": { "type": "string", "description": "Canonical single-line UTF-8 signature value; leading/trailing whitespace and control characters are rejected." },
          "header_protection_key": { "type": "string", "writeOnly": true, "maxLength": 44, "description": "Base64-encoded 32-byte HPK. Omit on PATCH to preserve; send an empty string to clear (link recreation required). All S1-S4 values must then be at least 12.", "example": "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=" },
          "header_protection_key_set": { "type": "boolean", "readOnly": true, "description": "Whether an HPK is configured; the key itself is never returned.", "example": true },
          "content_padding_addition": { "description": "Inclusive u16 content-padding scalar or range; zero uses the backend default.", "example": "10-20", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 65535 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "rekey_after_time": { "description": "Inclusive u16 rekey interval scalar or range in seconds; zero uses the backend default.", "example": "120-180", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 65535 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "rekey_timeout": { "description": "Inclusive u16 rekey timeout scalar or range in seconds; zero uses the backend default.", "example": "15-25", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 65535 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "reject_after_time": { "description": "Inclusive u16 reject-after interval scalar or range in seconds; zero uses the backend default.", "example": 90, "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 65535 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "keepalive_timeout": { "description": "Inclusive u16 transport keepalive-timeout scalar or range in seconds; distinct from peer PersistentKeepalive.", "example": "30-45", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 65535 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "max_handshake_attempts": { "description": "Inclusive u16 maximum-handshake-attempt scalar or range; zero uses the backend default.", "example": "4-8", "oneOf": [{ "type": "integer", "minimum": 0, "maximum": 65535 }, { "type": "string", "pattern": "^[0-9]+(-[0-9]+)?$" }] },
          "random_trailers": { "type": "boolean", "description": "Unsafe/client-specific opt-in; generated profiles leave this false." },
          "disable_cookies": { "type": "boolean", "description": "Unsafe opt-in that weakens DoS protection; generated profiles leave this false." }
        }
      },
      "IPv4Pools": {"type": "array", "minItems": 1, "maxItems": 16, "items": {"type": "string"}, "description": "Ordered canonical RFC1918 IPv4 CIDRs, /29 or larger. Each reserves network, first host (gateway) and broadcast. Allocation uses the first available pool; it never falls through to another interface. Overlap within this list, with any other interface (also disabled) or observed host routes is rejected. On create send either this field or ipv4_subnet; omit both for automatic selection. On PATCH the first pool must remain the primary and occupied pools cannot be removed/resized. Existing device addresses and keys remain stable.", "example": ["10.8.0.0/24", "10.8.1.0/24", "10.8.3.0/24"]},
...
      "Interface": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "listen_port": { "type": "integer", "description": "UDP port number, not a byte or time quantity." },
          "ipv4_subnet": { "type": "string", "description": "Immutable primary IPv4 CIDR; also the first item in ipv4_pools." },
          "ipv4_pools": { "$ref": "#/components/schemas/IPv4Pools" },
          "mtu": { "type": "integer", "description": "Tunnel IP-packet MTU in bytes (B); not a traffic allowance." },
          "public_key": { "type": "string" },
          "obfuscation": { "$ref": "#/components/schemas/Obfuscation" },
          "preset": { "type": "string", "enum": ["plain", "recommended", "performance", "balanced", "resilient", "suggested", "randomized", "custom"], "description": "Persisted profile classification; custom is inferred when explicit parameters are supplied." },
          "enabled": { "type": "boolean" },
          "backend_mode": { "type": "string", "enum": ["kernel", "userspace"] },
          "endpoint_override": { "type": "string", "description": "Canonical hostname or IP address, optionally with port." },
          "created_at": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." },
          "updated_at": { "type": "string", "format": "date-time", "description": "RFC3339 instant with an explicit timezone (for example 2026-10-30T12:00:00Z); responses use UTC (Z). Not a localized calendar date or Unix milliseconds." }
        }
      },
...
    }
  },
...
}